Keep track of and check state of connection, also complete user registration

+++ example-server.js
@@ -9,23 +9,35 @@ var fospServer = new Server(options);
 
 fospServer.on('connection', function(con) {
   log('Recieved a new connection: ' + con.id);
+
+  con.on('close', function() {
+    log('Closing connection: ' + con.id);
+  });
 });
 
 fospServer.on('connect', function(con, msg) {
-  if (msg.body.version === "0.1")
+  if (msg.body.version === "0.1") {
+    con.ctx.negotiated = true;
     con.sendSucceded(100, msg.seq);
-  else
-    con.sendFailed(400, msg.seq);
+    return;
+  }
+  con.sendFailed(400, msg.seq);
 });
 fospServer.on('authenticate', function(con, msg) {
+  if (!isNegotiated(con, msg))
+    return;
   db.authenticateUser(msg.body.name, msg.body.password, function(failed) {
-    if (failed)
+    if (failed) {
       con.sendFailed(402, msg.seq);
-    else
-      con.sendSucceded(210, msg.seq);
+      return;
+    }
+    con.ctx.authenticated = true;
+    con.sendSucceded(210, msg.seq);
   });
 });
 fospServer.on('register', function(con, msg) {
+  if (!isNegotiated(con, msg))
+    return;
   db.addUser(msg.body.name, msg.body.password, function(failed) {
     if (failed)
       con.sendFailed(500, msg.seq);
@@ -34,6 +46,8 @@ fospServer.on('register', function(con, msg) {
   });
 });
 fospServer.on('select', function(con, msg) {
+  if (!isAuthenticated(con, msg))
+    return;
   db.getNode(msg.uri.toString(), function(err, result) {
     if (err)
       con.sendFailed(500, msg.seq, {}, "Failed to retrieve data\n" + err);
@@ -44,6 +58,8 @@ fospServer.on('select', function(con, msg) {
   });
 });
 fospServer.on('create', function(con, msg) {
+  if (!isAuthenticated(con, msg))
+    return;
   db.setNode(msg.uri.toString(), msg.body, function(err, result) {
     if (err)
       con.sendFailed(500, msg.seq, {}, err);
@@ -52,6 +68,8 @@ fospServer.on('create', function(con, msg) {
   });
 });
 fospServer.on('update', function(con, msg) {
+  if (!isAuthenticated(con, msg))
+    return;
   db.updateNode(msg.uri.toString(), msg.body, function(err, result) {
     if (err)
       con.sendFailed(500, msg.seq, {}, err);
@@ -60,6 +78,8 @@ fospServer.on('update', function(con, msg) {
   });
 });
 fospServer.on('delete', function(con, msg) {
+  if (!isAuthenticated(con, msg))
+    return;
   db.deleteNode(msg.uri.toString(), function(err) {
     if (err)
       con.sendFailed(500, msg.seq);
@@ -68,6 +88,8 @@ fospServer.on('delete', function(con, msg) {
   });
 });
 fospServer.on('list', function(con, msg) {
+  if (!isAuthenticated(con, msg))
+    return;
   db.listChildren(msg.uri.toString(), function(err, children) {
     if (err)
       con.sendFailed(500, msg.seq);
@@ -75,6 +97,24 @@ fospServer.on('list', function(con, msg) {
       con.sendSucceded(200, msg.seq, {}, children);
   });
 });
+
+var isNegotiated = function(con, msg) {
+  if (con.ctx.negotiated)
+    return true;
+  log('Connection is not yet negotiated');
+  con.sendFailed(402, msg.seq);
+  con.close();
+  return false;
+}
+
+var isAuthenticated = function(con, msg) {
+  if (con.ctx.negotiated && con.ctx.authenticated)
+    return true;
+  log('Connection is not yet authenticated');
+  con.sendFailed(402, msg.seq);
+  con.close();
+  return false;
+}
 
 var log = function(text) {
   console.log("example-server: " + text);
